еще покрутил меню

+++ lib/measure.c
@@ -130,14 +130,14 @@ int getTemperature()
     return fabs(temp);
 }
 
-uint32_t getHandlePos()
+int8_t getHandlePos()
 {
     const int handleMin = 0;
     const int handleMax = 3400;
     int pos = ((((int)measures.handle) - handleMin) * 100) / (handleMax - handleMin);
     if( pos > 100 ) pos = 100;
     if( pos < 0 )   pos = 0;
-    return (uint32_t)pos;
+    return (int8_t)pos;
 }
 
 uint32_t getVoltage()
